decoupled plugin example
