+++ examples/basic/README.md
@@ -1,0 +1,14 @@
+Plugin Example
+--------------
+
+Compile this driver via:
+
+    go build .
+
+Compile the plugin itself via:
+
+    cd plugin/ && go build .
+
+You can then launch the plugin sample via:
+
+    ./basic
